feat: apply advanced upgrade card configs from Jules_Balancer

+++ go-server/game/upgrades_config.go
@@ -9,7 +9,7 @@ import (
 
 type UpgradeCardConfig struct {
 	ID       uint8
-	MaxLevel uint8
+	MaxLevel uint16
 	Type     string
 	Value    float64
 }
